Added CocoaPod build
