+++ Application/Sources/Initialization.swift
@@ -12,20 +12,6 @@ import NimbusRenderVideoKit
 import SwiftUI
 import UIKit
 
-@main
-class AppDelegate: UIResponder, UIApplicationDelegate {
-
-    func application(
-        _ application: UIApplication,
-        didFinishLaunchingWithOptions launchOptions: [UIApplication.LaunchOptionsKey: Any]?
-    ) -> Bool {
-        initializeNimbusSDK()
-        return true
-    }
-}
-
-
-
 public extension UIApplicationDelegate {
     
     func initializeNimbusSDK() {
