+++ Podfile
@@ -5,6 +5,8 @@ platform :ios, '15.0'
 target 'nimbus-ios-sample-pods' do
   use_frameworks!
 
+  pod 'Application', :path => 'Application'
+
   pod 'NimbusSDK', '2.8.1', subspecs: [
      'NimbusKit',               # Nimbus SDK
      'NimbusRenderStaticKit',   # Static Ad Renderer
